fix(runs): Run#temparature and Run#topP as double

+++ openai-core/src/commonMain/kotlin/com.aallam.openai.api/run/Run.kt
@@ -113,12 +113,12 @@ public data class Run(
     /**
      * The Unix timestamp (in seconds) for when the run was completed.
      */
-    @SerialName("temperature") val temperature: Int? = null,
+    @SerialName("temperature") val temperature: Double? = null,
 
     /**
      * The nucleus sampling value used for this run. If not set, defaults to 1.
      */
-    @SerialName("top_p") val topP: Int? = null,
+    @SerialName("top_p") val topP: Double? = null,
 
     /**
      * The maximum number of prompt tokens specified to have been used over the course of the run.
